// src/lib/model/enum/tauri-function.enum.ts
export enum TauriFunctionEnum {
	// Information
	GET_APP_VERSION = 'get_app_version',
	GET_IPV4_ADDRESS = 'get_ipv4_address',
	GET_CURRENT_ISO_TIME_STRING = 'get_current_iso_time_string',
	GET_CURRENT_YEAR_MONTH_DAY_STRING = 'get_current_year_month_day_string',
	GET_SYSTEM_INFORMATION = 'get_system_information',
	GET_DEVICE_INFORMATION = 'get_device_information',
	// Window Controller
	FOCUS_WINDOW = 'focus_window',
	HIDE_WINDOW = 'hide_window',
	MAXIMIZE_WINDOW = 'maximize_window',
	OPEN_WINDOW = 'open_window',
	REFRESH_WINDOW = 'refresh_window',
	SET_DEFAULT_SIZE_WINDOW = 'set_default_size_window',
	EXIT_WINDOW = 'exit_window',
	MINIMIZE_WINDOW = 'minimize_window'
}